5.7/requirements.txt
#기본 라이브러리
pandas==2.1.3
numpy==1.26.2
faiss-cpu==1.7.4
sentence-transformers==2.3.1
gtts==2.4.0
pygame==2.5.2
python-dotenv==1.0.0
PyQt5==5.15.10
#AI 관련 라이브러리
google-generativeai==0.3.1
#PDF 처리 라이브러리
PyMuPDF==1.23.7
#유틸리티 라이브러리
uuid==1.30
regex==2023.10.3
deepface
opencv-python
tf-keras
sounddevice
openai-whisper